+++ src/extension.ts
@@ -1,9 +1,9 @@
-import * as vscode     from 'vscode'
-import * as _file      from './CreateFile'
-import * as _test      from './CreateTest'
-import LensProvider    from './Lens/LensProvider'
+import * as vscode from 'vscode'
+import * as _file from './CreateFile'
+import * as _test from './CreateTest'
+import LensProvider from './Lens/LensProvider'
 import updateNamespace from './NamespaceUpdate'
-import * as utils      from './utils'
+import * as utils from './utils'
 
 export async function activate(context) {
     /* Other -------------------------------------------------------------------- */
@@ -35,12 +35,16 @@ export async function activate(context) {
         )
     }
 
-    vscode.workspace.onDidRenameFiles(async (event) => await updateNamespace(event))
+    vscode.workspace.onDidRenameFiles(async (event: vscode.FileRenameEvent) => await updateNamespace(event))
 }
 
 async function createFile(folder, type) {
     if (folder?.path) {
-        await _file.createFile(folder.path, type)
+        try {
+            await _file.createFile(folder.path, type)
+        } catch (error) {
+            return
+        }
     }
 
     await _file.insertSnippet(type)